+++ tests/e2e/test-box-lsp.R
@@ -1,0 +1,3 @@
+testthat::expect_false(
+  is.null(getOption("languageserver.parser_hooks"))
+)