+++ NEWS.md
@@ -1,6 +1,8 @@
 # rhino (development version)
 
 * Integrate `box.lsp` for auto-complete support for `box` modules in VS Code.
+* Integrated {box.linters} styling functions to style `box::use()` calls according to the Rhino style guide.
+* Added compatibility check for `treesitter` and `treesitter.r` dependencies
 
 # [rhino 1.9.0](https://github.com/Appsilon/rhino/releases/tag/v1.9.0)