+++ NEWS.md
@@ -1,5 +1,6 @@
 # rhino (development version)
 
+* Integrate `box.lsp` for auto-complete support for `box` modules in VS Code.
 * Integrated {box.linters} styling functions to style `box::use()` calls according to the Rhino style guide.
 * Added compatibility check for `treesitter` and `treesitter.r` dependencies